+++ test/model.test.js
@@ -6809,6 +6809,376 @@ describe('Model', function() {
       await Test.collection.drop().catch(() => {});
     });
 
+    it('createCollection() enforces expireAfterSeconds (gh-11229)', async function() {
+      this.timeout(10000);
+      const version = await start.mongodVersion();
+      if (version[0] < 5) {
+        this.skip();
+        return;
+      }
+
+      const schema = Schema({ name: String, timestamp: Date, metadata: Object }, {
+        timeseries: {
+          timeField: 'timestamp',
+          metaField: 'metadata',
+          granularity: 'hours'
+        },
+        autoCreate: false
+      });
+
+      const Test = db.model('TestGH11229Var1', schema);
+
+      await Test.collection.drop().catch(() => {});
+      await Test.createCollection({ expireAfterSeconds: 5 });
+
+      await Test.insertMany([
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T00:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T04:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T16:00:00.000Z'),
+          temp: 16
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T20:00:00.000Z'),
+          temp: 15
+        }, {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T00:00:00.000Z'),
+          temp: 13
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T04:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T16:00:00.000Z'),
+          temp: 17
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T20:00:00.000Z'),
+          temp: 12
+        }
+      ]);
+
+      const beforeExpirationCount = await Test.count({});
+      assert.ok(beforeExpirationCount === 12);
+      await new Promise(resolve => setTimeout(resolve, 6000));
+      const afterExpirationCount = await Test.count({});
+      assert.ok(afterExpirationCount === 0);
+      await Test.collection.drop().catch(() => {});
+    });
+
+    it('createCollection() enforces expires (gh-11229)', async function() {
+      this.timeout(10000);
+      const version = await start.mongodVersion();
+      if (version[0] < 5) {
+        this.skip();
+        return;
+      }
+
+      const schema = Schema({ name: String, timestamp: Date, metadata: Object }, {
+        timeseries: {
+          timeField: 'timestamp',
+          metaField: 'metadata',
+          granularity: 'hours'
+        },
+        autoCreate: false
+      });
+
+      const Test = db.model('TestGH11229Var2', schema);
+
+      await Test.collection.drop().catch(() => {});
+      await Test.createCollection({ expires: '5 seconds' });
+
+      await Test.insertMany([
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T00:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T04:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T16:00:00.000Z'),
+          temp: 16
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T20:00:00.000Z'),
+          temp: 15
+        }, {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T00:00:00.000Z'),
+          temp: 13
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T04:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T16:00:00.000Z'),
+          temp: 17
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T20:00:00.000Z'),
+          temp: 12
+        }
+      ]);
+
+      const beforeExpirationCount = await Test.count({});
+      assert.ok(beforeExpirationCount === 12);
+      await new Promise(resolve => setTimeout(resolve, 6000));
+      const afterExpirationCount = await Test.count({});
+      assert.ok(afterExpirationCount === 0);
+      await Test.collection.drop().catch(() => {});
+    });
+
+    it('createCollection() enforces expireAfterSeconds when set by Schema (gh-11229)', async function() {
+      this.timeout(10000);
+      const version = await start.mongodVersion();
+      if (version[0] < 5) {
+        this.skip();
+        return;
+      }
+
+      const schema = Schema({ name: String, timestamp: Date, metadata: Object }, {
+        timeseries: {
+          timeField: 'timestamp',
+          metaField: 'metadata',
+          granularity: 'hours'
+        },
+        autoCreate: false,
+        expireAfterSeconds: 5
+      });
+
+      const Test = db.model('TestGH11229Var3', schema);
+
+      await Test.collection.drop().catch(() => {});
+      await Test.createCollection();
+
+      await Test.insertMany([
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T00:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T04:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T16:00:00.000Z'),
+          temp: 16
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T20:00:00.000Z'),
+          temp: 15
+        }, {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T00:00:00.000Z'),
+          temp: 13
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T04:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T16:00:00.000Z'),
+          temp: 17
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T20:00:00.000Z'),
+          temp: 12
+        }
+      ]);
+
+      const beforeExpirationCount = await Test.count({});
+      assert.ok(beforeExpirationCount === 12);
+      await new Promise(resolve => setTimeout(resolve, 6000));
+      const afterExpirationCount = await Test.count({});
+      assert.ok(afterExpirationCount === 0);
+      await Test.collection.drop().catch(() => {});
+    });
+
+    it('createCollection() enforces expires when set by Schema (gh-11229)', async function() {
+      this.timeout(10000);
+      const version = await start.mongodVersion();
+      if (version[0] < 5) {
+        this.skip();
+        return;
+      }
+
+      const schema = Schema({ name: String, timestamp: Date, metadata: Object }, {
+        timeseries: {
+          timeField: 'timestamp',
+          metaField: 'metadata',
+          granularity: 'hours'
+        },
+        autoCreate: false,
+        expires: '5 seconds'
+      });
+
+      const Test = db.model('TestGH11229Var4', schema);
+
+      await Test.collection.drop().catch(() => {});
+      await Test.createCollection();
+
+      await Test.insertMany([
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T00:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T04:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T16:00:00.000Z'),
+          temp: 16
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-18T20:00:00.000Z'),
+          temp: 15
+        }, {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T00:00:00.000Z'),
+          temp: 13
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T04:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T08:00:00.000Z'),
+          temp: 11
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T12:00:00.000Z'),
+          temp: 12
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T16:00:00.000Z'),
+          temp: 17
+        },
+        {
+          metadata: { sensorId: 5578, type: 'temperature' },
+          timestamp: new Date('2021-05-19T20:00:00.000Z'),
+          temp: 12
+        }
+      ]);
+
+      const beforeExpirationCount = await Test.count({});
+      assert.ok(beforeExpirationCount === 12);
+      await new Promise(resolve => setTimeout(resolve, 6000));
+      const afterExpirationCount = await Test.count({});
+      assert.ok(afterExpirationCount === 0);
+      await Test.collection.drop().catch(() => {});
+    });
+
     it('createCollection() handles NamespaceExists errors (gh-9447)', async function() {
       const userSchema = new Schema({ name: String });
       const Model = db.model('User', userSchema);